VMS-29383: Rename LayoutTour entity to Showreel

+++ vms/client/nx_vms_client_desktop/src/nx/vms/client/desktop/ui/actions/action_types.h
@@ -115,7 +115,7 @@ enum ActionFlag
 
 
     /** Action can appear in main menu. */
-    Main = MainScope | NoTarget,
+    Main = MainScope | (int) NoTarget,
 
     /** Action can appear in scene context menu. */
     Scene = SceneScope | (int) WidgetTarget,
